+++ endpoint/http/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/endpoint/http/request/mapper/proprietary/PostOperationProviderByPathRequestMapper.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.ilt.faaast.service.endpoint.http.request.mapper.proprietary;
+
+import de.fraunhofer.iosb.ilt.faaast.service.ServiceContext;
+import de.fraunhofer.iosb.ilt.faaast.service.endpoint.http.model.HttpRequest;
+import de.fraunhofer.iosb.ilt.faaast.service.endpoint.http.request.mapper.AbstractSubmodelInterfaceRequestMapper;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.modifier.OutputModifier;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.request.proprietary.PostOperationProviderByPathRequest;
+import de.fraunhofer.iosb.ilt.faaast.service.model.api.response.proprietary.PostOperationProviderByPathResponse;
+import de.fraunhofer.iosb.ilt.faaast.service.model.exception.InvalidRequestException;
+import de.fraunhofer.iosb.ilt.faaast.service.model.http.HttpMethod;
+import de.fraunhofer.iosb.ilt.faaast.service.util.EncodingHelper;
+import de.fraunhofer.iosb.ilt.faaast.service.util.RegExHelper;
+import java.util.Map;
+
+
+/**
+ * class to map HTTP-POST-Request paths: submodels/{submodelIdentifier}/submodel-elements/{idShortPath},
+ * shells/{aasIdentifier}/submodels/{submodelIdentifier}/submodel-elements/{idShortPath}.
+ */
+public class PostOperationProviderByPathRequestMapper extends AbstractSubmodelInterfaceRequestMapper<PostOperationProviderByPathRequest, PostOperationProviderByPathResponse> {
+
+    private static final String OPERATION_PATH = RegExHelper.uniqueGroupName();
+    private static final String PATTERN = String.format("submodel-elements/%s/connection", pathElement(OPERATION_PATH));
+
+    public PostOperationProviderByPathRequestMapper(ServiceContext serviceContext) {
+        super(serviceContext, HttpMethod.POST, PATTERN);
+    }
+
+
+    @Override
+    public PostOperationProviderByPathRequest doParse(HttpRequest httpRequest, Map<String, String> urlParameters, OutputModifier outputModifier) throws InvalidRequestException {
+        // Can't parse body here, because submodelId is not yet available
+        return PostOperationProviderByPathRequest.builder()
+                .path(EncodingHelper.urlDecode(urlParameters.get(OPERATION_PATH)))
+                .body(httpRequest.getBodyAsString())
+                .build();
+    }
+
+}