+++ docs/source/other/release-notes.md
@@ -19,9 +19,10 @@
 - Endpoint
 	- HTTP
 		- URL prefix /api/v3.x is now optional
-		- Add new ocnfig property `httpVersion`
+		- Add new config property `httpVersion`
 		- Treat headers case-insenstive (see RFC 2616)
 		- OperationProvider now supports use of input arguments via variables in headers
+		- Add support for runtime update of OperationProvider
 
 **Internal changes & bugfixes**
 - General