+++ docs/source/other/release-notes.md
@@ -33,6 +33,7 @@
     - Fix ElementDelete events not being triggered on DELETE /reset.
 	- Removed duplicate requests that have not been mapped to any API (`GetAssetAdministrationShellByIdRequest`, `PutAssetAdministrationShellById`, `GetSubmodelByIdRequest`, `PatchSubmodelByIdRequest`, `PutSubmodelByIdRequest`)
 	- Update `ServiceContext` interface to now allow access to persistence, file storage, and asset connection manager
+    - Allow more flexibility in license header: Owners registered in CONTRIBUTORS.txt may add their Copyright statement to source files
 - Asset Connection
 	- Fixed bug that reading a SubmodelElement container, like a `SubmodelElementCollection`, didn't trigger the asset connection value providers of underlying elements recursively.
 	- Fixed asset connection JSON serialization not including blob values in serialization